[WB-2324.1] Reuse hash logic for vite and rollup
