+++ vite.config.ts
@@ -2,10 +2,21 @@ import {resolve} from "path";
 import react from "@vitejs/plugin-react-swc";
 import {defineConfig} from "vite";
 
+// Shared with Rollup so Storybook (Vite) and the published `dist` (Rollup)
+// produce byte-identical hashed CSS Modules class names from the same
+// `.module.css` source. Drift between the two would mean Storybook
+// previews reference selectors that don't exist in the shipped CSS.
+import {generateScopedName} from "./build-settings/css-modules-scoped-name.mjs";
+
 export default defineConfig({
     plugins: [react()],
     build: {
         assetsInlineLimit: 0,
+    },
+    css: {
+        modules: {
+            generateScopedName,
+        },
     },
     resolve: {
         alias: [
